+++ django_project/gap_dashboard/tests/views/dashboard/admin/__init__.py
@@ -1,2 +1,1 @@
-from .geography import *
 from .instance import *
