--- django_project/gap_dashboard/tests/views/dashboard/admin/geography/__init__.py
@@ -1,1 +1,0 @@
-from .level_management import *
