--- django_project/gap_dashboard/views/__init__.py
@@ -1,2 +1,0 @@
-__author__ = 'Irwan Fathurrahman <meomancer@gmail.com>'
-__date__ = '19/10/21'
