--- django_project/gap_dashboard/views/dashboard/admin/geography/__init__.py
@@ -1,3 +1,0 @@
-from .level_management import *
-from .upload import *
-from .view import *
