+++ django_project/gap_data/admin/__init__.py
@@ -1,6 +1,6 @@
 from .basemap_layer import *
 from .context_layer import *
 from .dashboard import *
-from .geometry import *
+from .reference_layer import *
 from .indicator import *
 from .link import *
