+++ django_project/gap_data/models/__init__.py
@@ -1,8 +1,6 @@
 from .basemap_layer import *
 from .context_layer import *
 from .dashboard import *
-from .geometry import *
-from .geometry_uploader import *
 from .indicator import *
 from .instance import *
 from .link import *
