+++ django_project/gap_data/tests/model_factories/__init__.py
@@ -1,7 +1,7 @@
 from .indicator import *  # noqa
 from .basemap_layer import *  # noqa
 from .context_layers import *  # noqa
-from .geometry import *  # noqa
+from .reference_layer import *  # noqa
 from .instance import *  # noqa
 from .link import *  # noqa
 from .user import *  # noqa
